Add fix for setting translatable fields language
I disovered that even though the forms save correctly, the actual
translated fields weren't getting saved. The issue lies in backdrop's
version of locale_field_node_form_submit which used to call
entity_language. entity_language allowed us to set a callback on the
entity to get the language from the entity translation handler. The
backdrop version of this code uses $form['#language'] which is set to
the default language and not the entity translation handler's code. The
function I added switches the language on translatable fields  back to
the entity translation handler's language.