Fix up alias handling
It took me a long time to figure out the proper way to do this. After
reviewing how the path module now sets the langcode and updates the
automatic aliases, it became clear that the easiest way to handle the
case for aliases on translations was to completely override the core
path module's implementation path_entity_update and to do everything
myself in entity_translations_entity_update. We fall back to core's
path_entity_update in the case when we're not handling a translation.

I tried messing around with the path_pattern_alter and path_alias_alter
but those were proving to be more difficult than the hook_entity_update
route.

There is still a pathauto hook implemented in here. I adjusted the
variable_get calls on the settings page, but I still need to rewrite the
code for bulk updating automatic aliases, I have an open issue for this
over in
https://github.com/backdrop-contrib/entity_translation/issues/1