Correct some compatibility with path
A hook inteneded for batch operations was refrencing the
'pathauto' module, which is now in core. I deleted it,
because it was not doing anything. Some relevent config
was also deleted because it was only used there.

Important to note: batch path operations are not implemented.
As far as I can tell, they have likely never worked for this module.
We will have to implement hook_path_info() to do this correctly.