Set node langcode in entity_translation_node_page_view
I don't know exactly where the langcode would get set in Drupal but in
Backdrop I think we have to set it here, otherwise if the page is
originally in a RTL language and we're viewing it in a LTR language it
will get shown in the wrong direction.

Also the page title was showing in the wrong language and I tracked it
down to using drupal_set_title instead of backdrop_set_title.